requirements.txt
fastapi==0.115.7
langchain==0.3.15
langchain-community==0.3.15
nltk==3.9.1
numpy==1.26.4
openai==0.28.0
pandas==2.2.3
python-dotenv==1.0.1
rouge_score==0.1.2
sacrebleu==2.5.1
sentence-transformers==3.4.0
streamlit==1.41.1
uvicorn==0.25.0
pydantic
python-multipart